third_party/WebKit/LayoutTests/editing/undo/undo-typing-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => {
        assert_selection(
            '<div contenteditable><span>|XX</span></div>',
            selection => {
                selection.document.execCommand('insertText', false, 'a');
                selection.document.execCommand('undo');
            },
            '<div contenteditable><span>|XX</span></div>');
    }, 'Undo typing a character');
</script>
